refactor: extract CQL into feature/cql (part of #778)
Move the CQL (Cassandra interface) statement family out of core
internal/mycli into internal/mycli/feature/cql behind the Feature
registration seam, so core no longer imports github.com/gocql/gocql or
github.com/googleapis/go-spanner-cassandra. This mirrors the BIGQUERY
extraction (#781).

- New package internal/mycli/feature/cql: CQLStatement, the fail-closed
  firstCQLKeyword/cqlStatementMutates classifier, formatCassandraTypeName,
  and Execute. The statement embeds mycli.MutationClassifier (Classify
  wired in newCQLStatement over its own CQL text) and carries its own
  compile-time ConditionallyMutatingStatement assertion. It is NOT
  detached-compatible (the Cassandra adapter binds to the Spanner
  database), so it does not mark MarksDetachedCompatible.
- Feature() contributes only the "CQL" statement def (EARLY EXPERIMENTAL);
  no vars, no flags, fresh instances per call, no package-level state.
- The lazy CQL session is reimplemented on mycli.FeatureState (key
  "cql.session") as a mutex-guarded, build-once io.Closer holder,
  preserving today's spancql.NewCluster/DatabaseUri/Timeout build and
  error messages. Session sheds its cqlCluster/cqlSession fields and the
  Close block; the core assertion for CQLStatement is removed.
- feature/all.All() now returns {cql.Feature(), bigquery.Feature()},
  keeping the merged order LLM, CQL, BIGQUERY so generated docs stay
  byte-identical.
- Tests: classifier and dispatch/classification tests move into
  feature/cql; the dispatch-level READONLY guard for mutating CQL moves
  into an external package mycli_test file (cql_readonly_guard_test.go).

core no longer imports gocql / go-spanner-cassandra
(verified via go list -deps ./internal/mycli); the root main still pulls
them through feature/all.

Part of #778
