+++ internal/mycli/readonly_guard_test.go
@@ -71,90 +71,13 @@ func TestReadOnlyGuardPreservesActiveBatch(t *testing.T) {
 	}
 }
 
-// TestCQLStatementMutates exhaustively checks the fail-closed classification of
-// CQL statements: only SELECT is treated as read-only; every other verb
-// (mutations, DDL, permissions) and any unrecognized or empty keyword is
-// treated as mutating.
-func TestCQLStatementMutates(t *testing.T) {
-	t.Parallel()
-
-	for _, tt := range []struct {
-		cql  string
-		want bool
-	}{
-		{cql: "SELECT * FROM t", want: false},
-		{cql: "select * from t", want: false},
-		{cql: "  SELECT * FROM t", want: false},
-		{cql: "INSERT INTO t (id) VALUES (1)", want: true},
-		{cql: "UPDATE t SET c = 1 WHERE id = 1", want: true},
-		{cql: "DELETE FROM t WHERE id = 1", want: true},
-		{cql: "BEGIN BATCH INSERT INTO t (id) VALUES (1) APPLY BATCH", want: true},
-		{cql: "TRUNCATE t", want: true},
-		{cql: "CREATE TABLE t (id int PRIMARY KEY)", want: true},
-		{cql: "DROP TABLE t", want: true},
-		{cql: "ALTER TABLE t ADD c int", want: true},
-		{cql: "GRANT SELECT ON t TO role", want: true},
-		{cql: "REVOKE SELECT ON t FROM role", want: true},
-		{cql: "BOGUS not a real verb", want: true}, // fail-closed on unknown keyword
-		{cql: "", want: true},                      // fail-closed on empty
-		{cql: "   ", want: true},                   // fail-closed on whitespace-only
-	} {
-		t.Run(tt.cql, func(t *testing.T) {
-			t.Parallel()
-			if got := cqlStatementMutates(tt.cql); got != tt.want {
-				t.Errorf("cqlStatementMutates(%q) = %v, want %v", tt.cql, got, tt.want)
-			}
-		})
-	}
-}
-
-// TestReadOnlyGuardBlocksMutatingCQL verifies that mutating CQL is rejected by
-// Session.ExecuteStatement in READONLY mode. These statements short-circuit at
-// the guard before CQLStatement.Execute, so no Cassandra adapter is needed.
-//
-// The complementary case (read-only SELECT is NOT blocked) is covered by
-// TestReadOnlyGuardAllowsReadOnlyCQL and TestCQLStatementMutates; it is not
-// exercised through ExecuteStatement here because a passing SELECT proceeds to
-// Execute, which would attempt a real Cassandra/Spanner connection.
-func TestReadOnlyGuardBlocksMutatingCQL(t *testing.T) {
-	t.Parallel()
-
-	for _, tt := range []struct {
-		desc string
-		cql  string
-	}{
-		{desc: "DELETE blocked", cql: "DELETE FROM t WHERE id = 1"},
-		{desc: "INSERT blocked", cql: "INSERT INTO t (id) VALUES (1)"},
-		{desc: "unrecognized keyword blocked", cql: "FROBNICATE t"},
-	} {
-		t.Run(tt.desc, func(t *testing.T) {
-			t.Parallel()
-			session := newSessionForLocalVarTest(t)
-			session.systemVariables.Transaction.ReadOnly = true
-
-			_, err := session.ExecuteStatement(t.Context(), &CQLStatement{CQL: tt.cql})
-			if !errors.Is(err, errReadOnly) {
-				t.Errorf("%s in READONLY mode: got error %v, want errReadOnly", tt.desc, err)
-			}
-		})
-	}
-}
-
-// TestReadOnlyGuardAllowsReadOnlyCQL verifies that a read-only CQL statement is
-// not classified as mutating, so the READONLY guard lets it through. The guard
-// input is isConditionallyMutating; asserting it directly avoids running
-// CQLStatement.Execute, which would need a live Cassandra adapter.
-func TestReadOnlyGuardAllowsReadOnlyCQL(t *testing.T) {
-	t.Parallel()
-
-	stmt := &CQLStatement{CQL: "SELECT * FROM t"}
-	if _, isMutation := any(stmt).(MutationStatement); isMutation {
-		t.Fatal("CQLStatement must not be a static MutationStatement")
-	}
-	if stmt.isConditionallyMutating() {
-		t.Error("read-only SELECT CQL classified as mutating; READONLY guard would wrongly block it")
-	}
-}
+// CQL READONLY guard coverage moved with the family to
+// internal/mycli/feature/cql (#778): the fail-closed classifier is unit tested
+// there (TestCQLStatementMutates), the wired read-only/mutating classification
+// is asserted directly in TestCQLStatementConditionalMutation, and the
+// dispatch-level guard for mutating CQL is proven through real dispatch in the
+// external mycli_test package (cql_readonly_guard_test.go), which can import the
+// feature package without an import cycle.
 
 // BIGQUERY READONLY guard coverage moved with the family to
 // internal/mycli/feature/bigquery (#778): the fail-closed classifier is unit
